Extend oas-router to accept explicitly / previously defined content-types
The current implementation does not properly work for endpoints that serve
multiple content-types like for example an file / image endpoint that
should be able to provide different types of images to be rendered or used
by the browser or other HTML clients. Therefore the logic for setting the
correct content-type should reside in the controller. However, the current
implementation overwrites any changes made to the content-type header and
always sets it to application/json by default. Therefore the following
changes have been made to take a previously set content-type, the accepted
and available validation types into consideration.